rtlwifi_new: rtl8723*: Reduce latency for rtl8723Xe
In routine rtl8723_phy_rf_serial_read(), there are three mdelay(1)
that can be replaced by a single udelay(120) in the line of the 3rd
one. In addition, the routine is no longer locked with
spin_lock_irqsave(), but with spin_lock(), thus interrupts are not
disabled. Both changes help latency in the system.

Signed-off-by: Larry Finger <Larry.Finger@lwfinger.net>